elbaas-requirements.txt
blinker==1.9.0;MIT;LICENSE
click==8.1.8;BSD-3-Clause;LICENSE
Flask==3.1.0;BSD-3-Clause;LICENSE
itsdangerous==2.2.0;BSD-3-Clause;LICENSE
Jinja2==3.1.5;BSD-3-Clause;LICENSE
MarkupSafe==3.0.2;BSD-3-Clause;LICENSE
pyserial==3.5;Python-2.0;LICENSE
Werkzeug==3.1.3;BSD-3-Clause;LICENSE

